+++ HISTORY.md
@@ -4,7 +4,9 @@
 
 ### API
 
-- Working sidebar and recurrences (#33)
+- Add sidebar in UI (#33)
+- Show extra description of the DP (#33)
+- Make `create_figure` return a `Figure` object (#34)
 - Add support for displaying multiple arrays (#23, #32)
 - Improve Front-end Callback Structure (#31)
 - Implement object-oriented API (#28)